requirements.txt
nose
sphinx
maxminddb
dnspython
ipaddress
